Refactor: 디렉토리명 변경
