+++ proma/src/main/java/ai/platform/proma/repository/ChatRoomRepository.java
@@ -1,4 +1,4 @@
-package ai.platform.proma.repositroy;
+package ai.platform.proma.repository;
 
 import ai.platform.proma.domain.ChatRoom;
 import ai.platform.proma.domain.User;
