+++ proma/src/main/java/ai/platform/proma/repository/CommunicationMethodRepository.java
@@ -1,4 +1,4 @@
-package ai.platform.proma.repositroy;
+package ai.platform.proma.repository;
 import ai.platform.proma.domain.Prompt;
 import ai.platform.proma.domain.PromptMethods;
 import ai.platform.proma.domain.enums.PromptMethod;
